+++ src/main/scala/org/codeprose/api/Api.scala
@@ -4,7 +4,12 @@ import java.io.File
 import scala.collection.mutable.ArrayBuffer
 import org.codeprose.util.DynamicPropertyMap
 
-
+/**
+ * Class to be exchanged between Provider <-> Broker <-> Consumer
+ * 
+ * Use language specific key-value-pairs to enrich the tokens and save 
+ * project summary information. 
+ */
 class ProjectInfo (  
   val enrichedTokens : ArrayBuffer[(java.io.File, ArrayBuffer[Token])],
   val summary : ProjectSummary
@@ -13,7 +18,17 @@ class ProjectInfo (
 
 
 /**
+ * Basis for languages to to supported by codeprose.
  * 
+ * Use to include: 
+ *  - Token keys
+ *  - Summary keys
+ *  - Token types
+ *  - Source symbols
+ *  - ...
+ *  
+ *  
+ *  See ScalaLang as an example.
  */
 trait DefaultLang {
 	import DynamicPropertyMap._
