--- src/main/scala/org/codeprose/api/common/TokenProperties.scala
@@ -1,5 +1,0 @@
-package org.codeprose.api.common
-
-class TokenProperties {
-  
-}
