+++ src/main/scala/org/codeprose/consumer/WriterLatex.scala
@@ -3,6 +3,9 @@ package org.codeprose.consumer
 import org.codeprose.api.ProjectInfo
 
 class WriterLatex extends Consumer {
+  
+  def initialize() : Unit = {}
+  def close() : Unit = {}
   def generateOutput(projectInfo: ProjectInfo) : Unit = {
     ???
   }
